Make sure basic_level and comments are written to the correct column
This commit should fix #2, though there are potentially a considerable
number of bugs still there. This program should probably be rewritten
to handle the upcoming new CLAN format anyway. One thing I noticed was
that comments should avoid having commas in them because they end up
being treated as delimiters when written to the csv file. This ends up
adding an extra column past I, when H should be the last column. This
isn't really a problem with parse_clan itself, but something to take
note of.